// src/generated/hotel-search.ts
export interface components {
    schemas: {
        HotelSearchRequest: {
            HotelSearchRQ: {
                POS?: components["schemas"]["POS"];
                SearchCriteria: components["schemas"]["SearchCriteria"];
            };
        };
        HotelSearchResponse: {
            HotelSearchRS?: {
                ApplicationResults?: components["schemas"]["ApplicationResults"];
                HotelSearchInfos?: components["schemas"]["HotelSearchInfos"];
            };
        };
// ...
        RefPoint: {
// ...
            StateProv?: string;
            /**
             * @description The Country code of the reference point.
             * @example US
             */
            CountryCode?: string;
        };
// ...
        ApplicationResults: {
            /**
             * @description Specifies if the status is complete/incomplete/not processed/unknown.
             * @example Complete
             * @enum {string}
             */
            status?: "Complete" | "Incomplete" | "NotProcessed" | "Unknown";
            Success?: components["schemas"]["ElementStructure"];
            Error?: components["schemas"]["ElementStructure"];
            Warning?: components["schemas"]["ElementStructure"];
        };
        /** @description Lists the results of the application's search, as well as any errors that may have occured. */
        ElementStructure: {
            /**
             * @description Specifies the source of error when processing the request.
             * @example Transport
             */
            type?: string;
            /**
             * @description Contains current application server time stamp.
             * @example 2024-06-09T20:45:04+05:30
             */
            timeStamp?: string;
            /** @description Lists the application response status. */
            SystemSpecificResults?: {
                /**
                 * @description Contains current application server time stamp.
                 * @example 2024-06-09T20:45:04+05:30
                 */
                timeStamp?: string;
                /**
                 * @description If present, provides the name of the field for which the results apply.
                 * @example reference
                 */
                reference?: string;
                /** @description Contains an abbreviated version of the error in text format. */
                ShortText?: Record<string, never>;
                /** @description Contains the identity of the unknown or misspelled tag that caused the processing error. */
                Element?: Record<string, never>;
                /** @description Contains the identity of the record that failed among a group of records. This value may contain a concatenation of a unique failed transaction ID with specific record(s) associated with that transaction. */
                RecordID?: Record<string, never>;
                /** @description Contains an online description of the error that occurred. */
                DocURL?: Record<string, never>;
                /** @description Contains the host system common run to create the result. */
                HostCommand?: {
                    /**
                     * @description The Line Number Interchange Address Terminal Address (LNIATA) number.
                     * @example AAVTYZ
                     */
                    LNIATA?: string;
                };
                /** @description Lists application-specific codes and messages. A textual description to provide more information about the specific condition, warning, or error with the code attribute as a numeric value. */
                Message?: {
                    /**
                     * @description The code, if returned, for an error or warning.
                     * @example ERR.0381
                     */
                    code?: string;
                    /**
                     * @description The message associated to the above code.
                     * @example Invalid check-in date
                     */
                    value?: string;
                }[];
            }[];
        }[];
        /** @description Contains hotel availability details for all properties found. */
        HotelSearchInfos: {
// ...
            HotelSearchInfo?: components["schemas"]["HotelSearchInfo"];
        };
        /** @description Lists basic hotel, rate, and image information. */
        HotelSearchInfo: {
            HotelInfo: components["schemas"]["HotelInfo"];
            HotelImageInfo?: components["schemas"]["HotelImageInfo"];
        }[];
        /** @description Contains basic hotel information. */
        HotelInfo: {
// ...
            LocationInfo?: components["schemas"]["LocationInfo"];
            Amenities?: components["schemas"]["Amenities"];
            SecurityFeatures?: components["schemas"]["SecurityFeatures"];
            PropertyTypeInfo?: components["schemas"]["PropertyTypeInfo"];
            PropertyQualityInfo?: components["schemas"]["PropertyQualityInfo"];
        };
        /** @description Contains agency or Supplier labels assigned to the property. */
        TierLabels: {
            TierLabel?: components["schemas"]["TierLabel"];
        };
// ...
        LocationInfo: {
            /**
             * @description Specifies the geographical latitude of the hotel location.
             * @example 36.154800
             */
            Latitude?: string;
            /**
             * @description Specifies the geographical longitude of the hotel location.
             * @example -95.990356
             */
            Longitude?: string;
            Address?: components["schemas"]["Address"];
            Neighborhoods?: components["schemas"]["Neighborhoods"];
            Contact?: components["schemas"]["Contact"];
        };
        /** @description Contains address information for a hotel. */
        Address: {
// ...
            CityName: components["schemas"]["CityName"];
            StateProv?: components["schemas"]["StateProv"];
            /**
             * @description The postal code information for the property.
             * @example 60173
             */
            PostalCode?: string;
            CountryName: components["schemas"]["CountryName"];
        };
// ...
        Neighborhoods: {
            Neighborhood?: components["schemas"]["Neighborhood"];
        };
// ...
        Amenities: {
            Amenity?: components["schemas"]["Amenity"];
        };
        Amenity: {
            /**
             * @description The OTA Code mapped to the HAC OTA Code.
             * @example 5
             */
            Code?: number;
            /**
             * @description A description of the OTA Code.
             * @example Parking
             */
            Description?: string;
            /**
             * @description If true, the amenity mentioned is free of charge.
             * @example false
             */
            ComplimentaryInd?: boolean;
            /**
             * @description Amenity value
             * @example Ask at reception for car park charges
             */
            value?: string;
        }[];
        SecurityFeatures: {
            SecurityFeature?: components["schemas"]["SecurityFeature"];
        };
// ...
        PropertyTypeInfo: {
            PropertyType?: components["schemas"]["PropertyType"];
        };
// ...
        PropertyQualityInfo: {
            PropertyQuality?: components["schemas"]["PropertyQuality"];
        };
// ...
        HotelImageInfo: {
            ImageItem?: components["schemas"]["ImageItem"];
        };
        /** @description Contains image information for a hotel. */
        ImageItem: {
// ...
            Image: components["schemas"]["Image"];
            Category: components["schemas"]["Category"];
            AdditionalInfo: components["schemas"]["AdditionalInfo"];
        };
// ...
        Description: {
            /** @description Lists descriptions in all requested language codes. */
            Text?: {
                /**
                 * @description The ISO 639-1 two-letter language code.
                 * @example en
                 */
                Language?: string;
                /**
                 * @description Language Code
                 * @example The text description of the image.
                 */
                value?: string;
            }[];
        };
        /** @description Contains any additional information related to the media item. */
        AdditionalInfo: {
            Info?: components["schemas"]["Info"];
        };
        /** @description Lists additional information, like caption, in the response. */
        Info: {
            /**
             * @description Identifies the type of additional information present.
             * @example CAPTION
             * @enum {string}
             */
            Type: "CAPTION";
            Description: components["schemas"]["Description"];
        }[];
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}